## What this PR does / why we need it:

It optimises the `tm_try()` funcall by declaring the function return type as a static cty.DynamicPseudoType. This is safe because Terramate doesn't do any schema validation using Hashicorp evaluator.

This change alone gives ~0.5s performance improvement when running `generate` in our internal framework project.

Depends on #1280 

```
➜  iac-gcloud git:(main) ✗ /usr/bin/time -h ~/src/terramate/bin/terramate generate
Nothing to do, generated code is up to date
	6.73s real		11.81s user		0.43s sys
➜  iac-gcloud git:(main) ✗ /usr/bin/time -h ~/src/terramate/bin/terramate generate
Nothing to do, generated code is up to date
	6.72s real		11.84s user		0.43s sys
➜  iac-gcloud git:(main) ✗ /usr/bin/time -h ~/src/terramate/bin/terramate generate
Nothing to do, generated code is up to date
	6.73s real		11.84s user		0.44s sys
```
compared to `main`:

```
➜  iac-gcloud git:(main) ✗ /usr/bin/time -h ~/src/terramate/bin/terramate generate
Nothing to do, generated code is up to date
	7.25s real		12.70s user		0.46s sys
➜  iac-gcloud git:(main) ✗ /usr/bin/time -h ~/src/terramate/bin/terramate generate
Nothing to do, generated code is up to date
	7.28s real		12.66s user		0.46s sys
➜  iac-gcloud git:(main) ✗ /usr/bin/time -h ~/src/terramate/bin/terramate generate
Nothing to do, generated code is up to date
	7.23s real		12.62s user		0.47s sys
```

## Which issue(s) this PR fixes:


## Special notes for your reviewer:

- The `try()` implementation was copied and only the `.Type` field was changed return an static DynamicPseudoType.

## Does this PR introduce a user-facing change?
```
no
```